@@ -33,7 +33,7 @@
     "cross-env": "^7.0.2",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.12.1",
-    "eslint": "^6.3.0",
+    "eslint": "^7.13.0",
     "eslint-config-react-app": "^5.2.0",
     "eslint-plugin-flowtype": "^4.6.0",
     "eslint-plugin-import": "^2.20.1",
@@ -52,7 +52,7 @@
     "piskel-cli": "^1.0.41",
     "prettier": "^1.13.7",
     "pretty-quick": "^2.0.1",
-    "react-scripts": "^3.4.3"
+    "react-scripts": "^4.0.0"
   },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.27",
